requirements.txt
Django==1.9
django-tastypie==0.14.0
docker==2.4.2
docker-compose==1.15.0
oauth2client==2.0.0
google-api-python-client==1.6.2
